+++ examples/singleton.js
@@ -1,5 +1,0 @@
-const {singleton: config} = require("../lib/index");
-
-module.exports = () => {
-    return config.envVars.bool;
-};