+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+# Version 2.0.0
+
+* Improved return type of compiled config
+* Removed singleton support
+* Exported default config builder for when you don't need to access loggable objects etc
+* Updated CI pipeline
+* NPM audit
+
 # Version 1.1.2
 
 * Rename module